change format mini photo

+++ src/main/java/it/sevenbits/FacultySite/web/service/gallery/ImageDescriptionService.java
@@ -22,7 +22,7 @@ public class ImageDescriptionService {
     @Autowired
     private ImageDescriptionRepository repository;
 
-    static public final double relationSide = ((long)16)/9;
+    static public final double relationSide = ((float)16)/9;
     static public final double miniImgWidth = 240;
     static public final double miniImgHeight = miniImgWidth/relationSide;
